[RED] Add test for upload invoice PDF response handling with authentication
